requirements.txt
httpx==0.27.0
PyJWT==2.9.0
alembic==1.13.2
Flask==3.0.3
Flask_JWT_Extended==4.6.0
flask_mail==0.10.0
Flask_Migrate==4.0.7
flask_restx==1.3.0
flask_sqlalchemy==3.1.1
pymysql==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.31
Werkzeug==3.1.3
openai==1.42.0
email-validator==2.2.0
flask-cors==5.0.1
gunicorn==23.0.0
gevent==24.11.1
fernet==1.0.1
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
msal==1.31.0
psycopg2-binary==2.9.10
redis==5.1.1
Flask-Limiter==3.8.0
BeautifulSoup4==4.12.3